controller/requirements.txt
bleak==0.22.3
pynput==1.7.7
paho-mqtt==2.1.0